added explicit int casts when printing GLint values, because some       older SDKs define GLint as long instead of int, e.g. Mac OSX SDKs       10.4 and older. removed static versions of osx codec import libs.       renamed osx codec dylib import lib directory to be lib.
git-svn-id: file:///svn/uhexen2/code/trunk@4915 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/h2shared/gl_vidsdl.c
@@ -663,7 +663,7 @@ static void CheckMultiTextureExtensions (void)
 			return;
 		}
 
-		Con_SafePrintf("Found %i TMUs support\n", num_tmus);
+		Con_SafePrintf("Found %i TMUs support\n", (int) num_tmus);
 		gl_mtexable = true;
 		glDisable_fp(GL_TEXTURE_2D);
 		glActiveTextureARB_fp(GL_TEXTURE0_ARB);
@@ -866,7 +866,7 @@ static void GL_Init (void)
 	glGetIntegerv_fp(GL_MAX_TEXTURE_SIZE, &gl_max_size);
 	if (gl_max_size < 256)	// Refuse to work when less than 256
 		Sys_Error ("hardware capable of min. 256k opengl texture size needed");
-	Con_SafePrintf("OpenGL max.texture size: %i\n", gl_max_size);
+	Con_SafePrintf("OpenGL max.texture size: %i\n", (int) gl_max_size);
 
 	is_3dfx = false;
 	if (!q_strncasecmp(gl_renderer, "3dfx", 4)	  ||
